Complete redesign of the scene model and introduction of the resource manager.
Materials are now stored as live Resources and are identified by unique path.
Scene is flat and consist of Nodes. Each represented by a 4x4 matrix(transform) and a shape.
Secondary array contain material ids mapped to the same index slots as the nodes.(Both are linear arrays)

Possibly overengineering the resource manager, better implement it carefully and when I truly need it.

Side note: The unique ids in the resource manager generally use 2 forms: string paths and their hashed variants. The string paths are the same for all instances, while hashes may be different for different workstations. If distributed ray tracing is about to be implemented, send accross the network a string path rather than hash.